refactor: extract service management into platform layer
Move macOS launchd and Linux systemd service implementations from
service.rs into platform/unix/service.rs. service.rs now delegates
and exposes dir_hash/service_label as shared utilities.
